Allow for construction of `Reader` from any type `R: Read + Seek`
This is a follow up to @est31's suggestion in #1 and is a follow on from
an earlier attempt at something similar I tried when first starting the
project.

The method attempts to detect the correct variant by attempting to
construct each of the format-specific readers and return the first one
that does not produce an error upon construction.

This feels a little hackish in that we depend on the upstream readers'
to first "confirm" the format and early return an `Err` (rather than
`panic!`ing or returning `Ok` anyway) if incorrect. Perhaps this is a
reasonable expectation though?

Also, this method requires that the successfully detected inner format
reader is constructed *twice*:

1. By passing the reader argument via `&mut` to test for the format. We
must use `&mut` to avoid moving the reader argument into the format
reader's constructor, otherwise we would not be able to re-use the
reader argument when testing the following formats.
2. Passing the reader argument via move for constructing the `Reader`
that will be returned after successfully detecting the format.

Perhaps a better/"proper" solution would be to dig a little deeper into
each of the upstream crates in order to find functionality specifically
suited to reading/confirming the header of each format? Or maybe it is
safe to assume that this would be the first step of the format reader
constructors anyway?

@est31 would be great to get your thoughts on all this!